+++ gcsio/src/test/java/com/google/cloud/hadoop/gcsio/GoogleCloudStorageTest.java
@@ -16,7 +16,6 @@ package com.google.cloud.hadoop.gcsio;
 
 import static com.google.cloud.hadoop.gcsio.GoogleCloudStorageImpl.createItemInfoForBucket;
 import static com.google.cloud.hadoop.gcsio.GoogleCloudStorageImpl.createItemInfoForStorageObject;
-import static com.google.cloud.hadoop.gcsio.GoogleCloudStorageImpl.initializeStorageRequestAuthorizer;
 import static com.google.cloud.hadoop.gcsio.GoogleCloudStorageItemInfo.createInferredDirectory;
 import static com.google.cloud.hadoop.gcsio.GoogleCloudStorageReadOptions.DEFAULT_BACKOFF_MAX_ELAPSED_TIME_MILLIS;
 import static com.google.cloud.hadoop.gcsio.GoogleCloudStorageTestUtils.HTTP_TRANSPORT;
@@ -69,8 +68,6 @@ import com.google.api.services.storage.model.Buckets;
 import com.google.api.services.storage.model.Objects;
 import com.google.api.services.storage.model.StorageObject;
 import com.google.cloud.hadoop.gcsio.GoogleCloudStorage.ListPage;
-import com.google.cloud.hadoop.gcsio.authorization.FakeAuthorizationHandler;
-import com.google.cloud.hadoop.gcsio.authorization.StorageRequestAuthorizer;
 import com.google.cloud.hadoop.util.AccessBoundary;
 import com.google.cloud.hadoop.util.ApiErrorExtractor;
 import com.google.cloud.hadoop.util.AsyncWriteChannelOptions;
@@ -3400,23 +3397,6 @@ public class GoogleCloudStorageTest {
             getRequestString(BUCKET_NAME, OBJECT_NAME),
             getRequestString(BUCKET_NAME, OBJECT_NAME))
         .inOrder();
-  }
-
-  @Test
-  public void testInstantiateAuthorizationHandler() {
-    GoogleCloudStorageOptions options =
-        GoogleCloudStorageOptions.builder()
-            .setAppName("test-name")
-            .setProjectId("test-project")
-            .setAuthorizationHandlerImplClass(FakeAuthorizationHandler.class)
-            .setAuthorizationHandlerProperties(
-                ImmutableMap.of(
-                    FakeAuthorizationHandler.PROPERTY_KEY, FakeAuthorizationHandler.EXPECTED_VALUE))
-            .build();
-
-    StorageRequestAuthorizer authorizer = initializeStorageRequestAuthorizer(options);
-
-    assertThat(authorizer).isNotNull();
   }
 
   @Test